+++ src/model/Model_Budget.h
@@ -72,7 +72,7 @@ public:
     static PERIOD_ID period_id(const Data& r);
     static DB_Table_BUDGETTABLE_V1::PERIOD PERIOD(PERIOD_ID period, OP op = EQUAL);
 
-    static void getBudgetEntry(int budgetYearID,
+    static void getBudgetEntry(int64 budgetYearID,
         std::map<int, PERIOD_ID> &budgetPeriod,
         std::map<int, double> &budgetAmt,
         std::map<int, wxString> &budgetNotes);
@@ -80,7 +80,7 @@ public:
         std::map<int, std::map<int, double> > &budgetStats
         , mmDateRange* date_range
         , bool groupByMonth);
-    static void copyBudgetYear(int newYearID, int baseYearID);
+    static void copyBudgetYear(int64 newYearID, int64 baseYearID);
     static double getEstimate(bool is_monthly, const PERIOD_ID period, const double amount);
 };
